requirements.txt
jinja2>=3.1.0
pyyaml>=6.0
jsonschema>=4.0.0
